+++ packages/framer-motion/src/animation/animators/waapi/__tests__/animate-style.test.ts
@@ -26,30 +26,6 @@ describe("animateMini", () => {
         )
     })
 
-    test("Applies target keyframe when animation has finished", async () => {
-        const div = document.createElement("div")
-        const animation = animateMini(
-            div,
-            { opacity: 0.6 },
-            { duration, x: {}, "--css-var": {} }
-        )
-        await animation.then(() => {
-            expect(div).toHaveStyle("opacity: 0.6")
-        })
-    })
-
-    test("Applies final target keyframe when animation has finished", async () => {
-        const div = document.createElement("div")
-        const animation = animateMini(
-            div,
-            { opacity: [0.2, 0.5] },
-            { duration }
-        )
-        await animation.then(() => {
-            expect(div).toHaveStyle("opacity: 0.5")
-        })
-    })
-
     test("time sets and gets time", async () => {
         const div = document.createElement("div")
         const animation = animateMini(div, { opacity: 0.5 }, { duration: 10 })
@@ -57,49 +33,6 @@ describe("animateMini", () => {
         expect(animation.time).toBe(0)
         animation.time = 5
         expect(animation.time).toBe(5)
-    })
-
-    test("autoplay false pauses animation", async () => {
-        const div = document.createElement("div")
-        const animation = animateMini(
-            div,
-            { opacity: 0.5 },
-            { duration: 0.1, autoplay: false }
-        )
-        let hasFinished = false
-
-        animation.then(() => {
-            hasFinished = true
-        })
-
-        await new Promise<void>((resolve) => {
-            setTimeout(() => {
-                expect(hasFinished).toBe(false)
-                resolve()
-            }, 200)
-        })
-    })
-
-    test("time can be set to duration", async () => {
-        const div = document.createElement("div")
-        div.style.opacity = "0"
-        const animation = animateMini(div, { opacity: 0.5 }, { duration: 1 })
-        animation.pause()
-        animation.time = 1
-
-        return new Promise<void>((resolve) => {
-            setTimeout(() => {
-                expect(div).toHaveStyle("opacity: 0.5")
-                resolve()
-            }, 50)
-        })
-    })
-
-    test("duration gets the duration of the animation", async () => {
-        const div = document.createElement("div")
-        const animation = animateMini(div, { opacity: 0.5 }, { duration: 10 })
-
-        expect(animation.duration).toBe(10)
     })
 
     test("Interrupt polyfilled transforms", async () => {