+++ packages/motion-dom/src/animation/utils/get-value-transition.ts
@@ -1,7 +1,7 @@
 export function getValueTransition(transition: any, key: string) {
-    return transition
-        ? transition[key as keyof typeof transition] ||
-              (transition as any)["default"] ||
-              transition
-        : undefined
+    return (
+        transition?.[key as keyof typeof transition] ??
+        transition?.["default"] ??
+        transition
+    )
 }